move geometry tests into directory
